Fixed broken stl binary parser and refactored folder structure
