--- include/meshlib/writers/obj.h
@@ -1,2 +1,0 @@
-
-#pragma once
