+++ resource/ASCII-box.stl
@@ -1,19 +1,86 @@
-
-
-
 solid Exported from Blender-2.93.5
-facet normal 1.111111 2.000000 3.000000
+facet normal 0.000000 0.000000 1.000000
 outer loop
-vertex 2.000000 3.000000 4.000000
-vertex -1.0340000 1.000000 1.000000
-vertex -1.060000 -1.000000 1.000000
+vertex 1.000000 1.000000 1.000000
+vertex -1.000000 1.000000 1.000000
+vertex -1.000000 -1.000000 1.000000
 endloop
 endfacet
-facet normal 1.0 0.000000 1.000000
+facet normal 0.000000 0.000000 1.000000
 outer loop
 vertex 1.000000 1.000000 1.000000
 vertex -1.000000 -1.000000 1.000000
 vertex 1.000000 -1.000000 1.000000
+endloop
+endfacet
+facet normal 0.000000 -1.000000 0.000000
+outer loop
+vertex 1.000000 -1.000000 -1.000000
+vertex 1.000000 -1.000000 1.000000
+vertex -1.000000 -1.000000 1.000000
+endloop
+endfacet
+facet normal 0.000000 -1.000000 0.000000
+outer loop
+vertex 1.000000 -1.000000 -1.000000
+vertex -1.000000 -1.000000 1.000000
+vertex -1.000000 -1.000000 -1.000000
+endloop
+endfacet
+facet normal -1.000000 0.000000 0.000000
+outer loop
+vertex -1.000000 -1.000000 -1.000000
+vertex -1.000000 -1.000000 1.000000
+vertex -1.000000 1.000000 1.000000
+endloop
+endfacet
+facet normal -1.000000 0.000000 0.000000
+outer loop
+vertex -1.000000 -1.000000 -1.000000
+vertex -1.000000 1.000000 1.000000
+vertex -1.000000 1.000000 -1.000000
+endloop
+endfacet
+facet normal 0.000000 0.000000 -1.000000
+outer loop
+vertex -1.000000 1.000000 -1.000000
+vertex 1.000000 1.000000 -1.000000
+vertex 1.000000 -1.000000 -1.000000
+endloop
+endfacet
+facet normal 0.000000 0.000000 -1.000000
+outer loop
+vertex -1.000000 1.000000 -1.000000
+vertex 1.000000 -1.000000 -1.000000
+vertex -1.000000 -1.000000 -1.000000
+endloop
+endfacet
+facet normal 1.000000 0.000000 0.000000
+outer loop
+vertex 1.000000 1.000000 -1.000000
+vertex 1.000000 1.000000 1.000000
+vertex 1.000000 -1.000000 1.000000
+endloop
+endfacet
+facet normal 1.000000 0.000000 0.000000
+outer loop
+vertex 1.000000 1.000000 -1.000000
+vertex 1.000000 -1.000000 1.000000
+vertex 1.000000 -1.000000 -1.000000
+endloop
+endfacet
+facet normal 0.000000 1.000000 0.000000
+outer loop
+vertex -1.000000 1.000000 -1.000000
+vertex -1.000000 1.000000 1.000000
+vertex 1.000000 1.000000 1.000000
+endloop
+endfacet
+facet normal 0.000000 1.000000 0.000000
+outer loop
+vertex -1.000000 1.000000 -1.000000
+vertex 1.000000 1.000000 1.000000
+vertex 1.000000 1.000000 -1.000000
 endloop
 endfacet
 endsolid Exported from Blender-2.93.5
