+++ source/main.c
@@ -28,9 +28,9 @@ TODO:
 
 #include <disk_manager/file_reader.h>
 #include <disk_manager/file_writer.h>
-#include <meshlib/readers/stl/ascii.h>
-#include <meshlib/readers/stl/binary.h>
 
+#include <meshlib/stl/readers/ascii.h>
+#include <meshlib/stl/readers/binary.h>
 
 #include <string.h>
 
@@ -66,34 +66,11 @@ int main(int argc, char** argv)
 		buf_get_at(binaryBox.user_data, i, &normal);
 		printf("\tnormal: (%f, %f, %f)\n", normal.x, normal.y, normal.z);	
 	}
-
-	write_binary_to_file("Copy-Binary-box.stl", stl_binary->bytes, stl_binary->element_count);
+	
 	buf_free(stl_binary);
 	buf_free(stl_data);
 	mesh_destroy(asciiBox);
 	mesh_destroy(binaryBox);
-	// mesh_t mesh = mesh_new(vertex3d_t);
-	// vertex3d_t vertices[2] = 
-	// {
-	// 	{ { 1, 1, 1, 1 }, { 0, 0, 0 }, { 0, 1, 0 }, { 1, 0 } },
-	// 	{ { 1, 1, 1, 1 }, { 0, 0, 0 }, { 0, 1, 0 }, { 1, 0 } }
-	// };
-	// mesh_vertex_add(mesh, &vertices[0]);
-	// mesh_vertex_addv(mesh, vertices, 2);
-	// vertex3d_t* __vertices = mesh_vertices(vertex3d_t, mesh);
-
-	// for(int i = 0; i < buf_get_element_count(mesh.vertices); i++)
-	// {
-	// 	vertex3d_t vertex = __vertices[i];
-	// 	log_msg("vertex[%u]: color(%.2f, %.2f, %.2f, %.2f), position(%.2f, %.2f, %.2f), normal(%.2f, %.2f, %.2f), uv(%.2f, %.2f)\n",
-	// 		i,
-	// 		vertex.color.x, vertex.color.y, vertex.color.z, vertex.color.w,
-	// 		vertex.position.x, vertex.position.y, vertex.position.z,
-	// 		vertex.normal.x, vertex.normal.y, vertex.normal.z,
-	// 		vertex.uv.x, vertex.uv.y);
-	// }
-	// mesh_vertices_clear(mesh);
-	// mesh_destroy(mesh);
 	safe_memory_terminate();
 	return 0;
 }
