+++ source/stl/ascii.c
@@ -1,49 +1,11 @@
 
-#include <meshlib/readers/stl/ascii.h>
-#include <meshlib/readers/stl/vertex.h>
+#include <meshlib/stl/readers/ascii.h>
+#include <meshlib/stl/parse_error.h>
+#include <meshlib/stl/stl.h>
 #include <meshlib/parsers/string.h>
 #include <meshlib/assert.h>
 #include <hpml/vec3/header_config.h>
 #include <hpml/vec3/vec3.h>
-
-enum STL_PARSE_ERROR
-{
-	STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND,
-	STL_PARSE_ERROR_FACET_NOT_FOUND,
-	STL_PARSE_ERROR_URECOGNIZED_KEYWORD,
-	STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT,
-	STL_PARSE_ERROR_END_LOOP_NOT_FOUND,
-	STL_PARSE_ERROR_END_FACET_NOT_FOUND,
-	STL_PARSE_ERROR_NORMAL_NOT_FOUND,
-	STL_PARSE_ERROR_OUTER_NOT_FOUND,
-	STL_PARSE_ERROR_LOOP_NOT_FOUND,
-	STL_PARSE_ERROR_VERTEX_NOT_FOUND,
-	STL_PARSE_ERROR_END_SOLID_NOT_FOUND
-};
-
-function_signature(static void, stl_parse_error, u64 error_type, u64 line_no);
-#define stl_parse_error(...) define_alias_function_macro(stl_parse_error, __VA_ARGS__)
-
-function_signature(static void, stl_parse_error, u64 error_type, u64 line_no)
-{
-	CALLTRACE_BEGIN();
-	switch(error_type) 
-	{
-		case STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_FACET_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_FACET_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_URECOGNIZED_KEYWORD: LOG_FETAL_ERR("STL_PARSE_ERROR_URECOGNIZED_KEYWORD\n"); break;
-		case STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT: LOG_FETAL_ERR("STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT\n"); break;
-		case STL_PARSE_ERROR_END_LOOP_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_LOOP_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_END_FACET_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_FACET_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_NORMAL_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_NORMAL_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_OUTER_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_OUTER_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_LOOP_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_LOOP_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_VERTEX_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_VERTEX_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_END_SOLID_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_SOLID_NOT_FOUND\n"); break;
-		default: LOG_FETAL_ERR("STL_PARSE_ERROR: Unkown Error\n");
-	}
-	CALLTRACE_END();
-}
 
 static void free_normals(void* buffer)
 {
@@ -64,9 +26,9 @@ function_signature(mesh_t, stl_parse_ascii, const char* text)
 
 	if(!string_parser_strcmp("solid"))
 		stl_parse_error(STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND, string_parser_line_count());
-	
-	string_parser_next_line();
-	string_parser_skip_whitespaces();
+
+	string_parser_next_line();	
+	string_parser_skip_any_whitespace();	
 
 	while(!string_parser_strcmp("endsolid"))
 	{	
@@ -79,8 +41,7 @@ function_signature(mesh_t, stl_parse_ascii, const char* text)
 		vec3_t(float) normal =  { string_parser_float(), string_parser_float(), string_parser_float() };
 		buf_push(normals, &normal);
 	
-		string_parser_next_line();
-		string_parser_skip_whitespaces();
+		string_parser_skip_any_whitespace();
 		
 		if(!string_parser_strcmp("outer"))
 			stl_parse_error(STL_PARSE_ERROR_OUTER_NOT_FOUND, string_parser_line_count());
@@ -90,8 +51,7 @@ function_signature(mesh_t, stl_parse_ascii, const char* text)
 
 		for(int i = 0; i < 3; i++)
 		{
-			string_parser_next_line();
-			string_parser_skip_whitespaces();
+			string_parser_skip_any_whitespace();
 
 			if(!string_parser_strcmp("vertex"))
 				stl_parse_error(STL_PARSE_ERROR_VERTEX_NOT_FOUND, string_parser_line_count());
@@ -103,22 +63,20 @@ function_signature(mesh_t, stl_parse_ascii, const char* text)
 			buf_push(vertices, &vertex);
 		}
 
-		string_parser_next_line(); 
-		string_parser_skip_whitespaces(); 
+		string_parser_skip_any_whitespace();
 
 		if(!string_parser_strcmp("endloop"))
 			stl_parse_error(STL_PARSE_ERROR_END_LOOP_NOT_FOUND, string_parser_line_count());
 
-		string_parser_next_line(); 
-		string_parser_skip_whitespaces();
+		string_parser_skip_any_whitespace();
 
 		if(!string_parser_strcmp("endfacet"))
 			stl_parse_error(STL_PARSE_ERROR_END_FACET_NOT_FOUND, string_parser_line_count());
 
-		string_parser_next_line();
-		string_parser_skip_whitespaces();
+		string_parser_skip_any_whitespace();
 	}
 
 	CALLTRACE_END();
 	return mesh;
 }
+
