+++ source/stl/binary.c
@@ -1,50 +1,11 @@
 
-#include <meshlib/readers/stl/binary.h>
-#include <meshlib/readers/stl/vertex.h>
+#include <meshlib/stl/readers/binary.h>
+#include <meshlib/stl/parse_error.h>
+#include <meshlib/stl/stl.h>
 #include <meshlib/parsers/binary.h>
 #include <meshlib/assert.h>
 #include <hpml/vec3/header_config.h>
 #include <hpml/vec3/vec3.h>
-
-enum STL_PARSE_ERROR
-{
-	STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND,
-	STL_PARSE_ERROR_FACET_NOT_FOUND,
-	STL_PARSE_ERROR_URECOGNIZED_KEYWORD,
-	STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT,
-	STL_PARSE_ERROR_END_LOOP_NOT_FOUND,
-	STL_PARSE_ERROR_END_FACET_NOT_FOUND,
-	STL_PARSE_ERROR_NORMAL_NOT_FOUND,
-	STL_PARSE_ERROR_OUTER_NOT_FOUND,
-	STL_PARSE_ERROR_LOOP_NOT_FOUND,
-	STL_PARSE_ERROR_VERTEX_NOT_FOUND,
-	STL_PARSE_ERROR_END_SOLID_NOT_FOUND
-};
-
-function_signature(static void, stl_parse_error, u64 error_type, u64 line_no);
-#define stl_parse_error(...) define_alias_function_macro(stl_parse_error, __VA_ARGS__)
-
-function_signature(static void, stl_parse_error, u64 error_type, u64 line_no)
-{
-	CALLTRACE_BEGIN();
-	switch(error_type) 
-	{
-		case STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_SOLID_HEADER_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_FACET_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_FACET_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_URECOGNIZED_KEYWORD: LOG_FETAL_ERR("STL_PARSE_ERROR_URECOGNIZED_KEYWORD\n"); break;
-		case STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT: LOG_FETAL_ERR("STL_PARSE_ERROR_INVALID_SINGLE_PRECISION_FLOAT\n"); break;
-		case STL_PARSE_ERROR_END_LOOP_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_LOOP_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_END_FACET_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_FACET_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_NORMAL_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_NORMAL_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_OUTER_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_OUTER_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_LOOP_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_LOOP_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_VERTEX_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_VERTEX_NOT_FOUND\n"); break;
-		case STL_PARSE_ERROR_END_SOLID_NOT_FOUND: LOG_FETAL_ERR("STL_PARSE_ERROR_END_SOLID_NOT_FOUND\n"); break;
-		default: LOG_FETAL_ERR("STL_PARSE_ERROR: Unkown Error\n");
-	}
-	CALLTRACE_END();
-}
-
 
 static void free_normals(void* buffer)
 {
@@ -69,8 +30,8 @@ function_signature(mesh_t, stl_parse_binary, const char* bytes, u64 length)
 		{
 			vec3_t(float) vertex = { binary_parser_float(), binary_parser_float(), binary_parser_float() };
 			buf_push(vertices, &vertex);
-			binary_parser_skip_bytes(sizeof(u16));
 		}
+		binary_parser_skip_bytes(sizeof(u16));
 		--triangle_count;
 	}
 	return mesh;
